--- client/src/components/Background/style.css
@@ -1,39 +1,0 @@
-.background {
-  align-items: flex-start;
-  display: flex;
-  height: 857px;
-  left: 83px;
-  overflow: hidden;
-  position: absolute;
-  top: 0px;
-  width: 1440px;
-}
-.bg {
-  background-color: var(--zircon);
-  height: 752px;
-  left: 152px;
-  position: absolute;
-  top: 37px;
-  width: 1440px;
-}
-.overlap-group1 {
-  height: 873px;
-  margin-left: -152px;
-  margin-top: -37px;
-  position: relative;
-  width: 1832px;
-}
-.purple-circle {
-  height: 406px;
-  left: 1239px;
-  position: absolute;
-  top: 0px;
-  width: 406px;
-}
-.turquoise-circle {
-  height: 168px;
-  left: 981px;
-  position: absolute;
-  top: 501px;
-  width: 168px;
-}
