--- client/src/components/Footer/index.js
@@ -1,18 +1,0 @@
-import React from "react";
-import "./style.css";
-
-function Footer(props) {
-  const { untitledDesign91 } = props;
-
-  return (
-    <div className="footer-mini-6">
-      <img
-        className="untitled-design-9-1"
-        alt="esl icon"
-        src={untitledDesign91}
-      />
-    </div>
-  );
-}
-
-export default Footer;
