--- client/src/components/GetStartedButton/style.css
@@ -1,29 +1,0 @@
-@import url("https://fonts.googleapis.com/css?family=Manrope:800,700,400");
-.adjustable-primary-large {
-  align-items: center;
-  background-color: var(--electric-violet);
-  border-radius: 8px;
-  display: flex;
-  height: 62px;
-  overflow: hidden;
-  padding: 0 39px;
-  width: 196px;
-}
-.label-i541147 {
-  letter-spacing: 0px;
-  line-height: 30px;
-  min-height: 30px;
-  min-width: 116px;
-  text-align: center;
-  white-space: nowrap;
-}
-.manrope-bold-white-20px {
-  color: white;
-  font-family: "Manrope", Helvetica;
-  font-size: 20px;
-  font-style: normal;
-  font-weight: 700;
-}
-:root {
-  --electric-violet: rgba(140, 48, 245, 1);
-}
