--- client/src/components/HeroSubText/style.css
@@ -1,28 +1,0 @@
-@import url("https://fonts.googleapis.com/css?family=Manrope:700,800,400");
-.manrope-extra-bold-white-28px {
-  color: transparent;
-  font-family: "Manrope", Helvetica;
-  font-size: 28px;
-  font-style: normal;
-  font-weight: 800;
-}
-.span {
-  color: rgba(0, 0, 0, 0.44);
-  font-family: "Manrope", Helvetica;
-  font-size: 28px;
-  font-weight: 800;
-  line-height: 28px;
-}
-.span1 {
-  color: rgba(0, 0, 0, 1);
-  font-family: "Manrope", Helvetica;
-  font-size: 28px;
-  font-weight: 800;
-  line-height: 28px;
-}
-.text-1 {
-  letter-spacing: 0px;
-  line-height: 40px;
-  min-height: 150px;
-  width: 517px;
-}
