--- client/src/components/HeroText/style.css
@@ -1,17 +1,0 @@
-@import url("https://fonts.googleapis.com/css?family=Manrope:700,800,400");
-.manrope-extra-bold-eerie-black-72px {
-  color: var(--eerie-black);
-  font-family: "Manrope", Helvetica;
-  font-size: 72px;
-  font-style: normal;
-  font-weight: 800;
-}
-.text-1 {
-  letter-spacing: 0px;
-  line-height: 98px;
-  min-height: 196px;
-  width: 732px;
-}
-:root {
-  --eerie-black: rgba(24, 25, 31, 1);
-}
