--- client/src/components/Navbar/style.css
@@ -1,102 +1,0 @@
-.essential-icons-chevron-down {
-  align-items: flex-end;
-  display: flex;
-  height: 18px;
-  margin-left: 2px;
-  padding: 6px 3px;
-  width: 18px;
-}
-.header-7 {
-  align-items: center;
-  display: flex;
-  height: 72px;
-  justify-content: flex-end;
-  left: 0px;
-  min-width: 821px;
-  padding: 10px 4px;
-  position: absolute;
-  top: 10px;
-}
-.header-menu-dropdown {
-  align-items: flex-end;
-  display: flex;
-  height: 20px;
-  margin-left: 30px;
-  width: 57px;
-}
-.label-i551357 {
-  letter-spacing: 0px;
-  line-height: 20px;
-  min-height: 20px;
-  min-width: 37px;
-  white-space: nowrap;
-}
-.label-i551365 {
-  letter-spacing: 0px;
-  line-height: 20px;
-  margin-left: 30px;
-  min-height: 20px;
-  min-width: 33px;
-  white-space: nowrap;
-}
-.label-i551367 {
-  letter-spacing: 0px;
-  line-height: 20px;
-  min-height: 20px;
-  min-width: 44px;
-  white-space: nowrap;
-}
-.label-i551646 {
-  letter-spacing: 0px;
-  line-height: 20px;
-  margin-left: 30px;
-  min-height: 20px;
-  min-width: 62px;
-  white-space: nowrap;
-}
-.menu-item-default {
-  height: 20px;
-  margin-left: 32px;
-  width: 63px;
-}
-.nav-items {
-  align-items: flex-start;
-  display: flex;
-  height: 20px;
-  left: 0px;
-  position: absolute;
-  top: 10px;
-  width: 381px;
-}
-.overlap-group4 {
-  height: 40px;
-  margin-left: 90px;
-  position: relative;
-  width: 381px;
-}
-.right-nav {
-  height: 40px;
-  left: 43px;
-  position: absolute;
-  top: 0px;
-  width: 135px;
-}
-.text-2 {
-  letter-spacing: 0px;
-  line-height: 30px;
-  margin-bottom: 4px;
-  margin-left: 22px;
-  min-height: 26px;
-  white-space: nowrap;
-  width: 114px;
-}
-.untitled-design-8-12 {
-  align-self: flex-end;
-  height: 44px;
-  object-fit: cover;
-  width: 45px;
-}
-.vector {
-  height: 6px;
-  width: 12px;
-}
