--- client/src/components/PreviewCard/style.css
@@ -1,32 +1,0 @@
-.content {
-  align-items: flex-start;
-  display: flex;
-  width: 111px;
-}
-.feedback-person-details {
-  height: 24px;
-  width: 111px;
-}
-.line-1 {
-  height: 1px;
-  margin-top: 33px;
-  width: 589px;
-}
-.namedentityextraction-1 {
-  height: 301px;
-  margin-top: 33px;
-  object-fit: cover;
-  width: 589px;
-}
-.preview-card {
-  align-items: flex-start;
-  background-color: white;
-  border-radius: 8px;
-  box-shadow: 0px 10px 20px rgba(41, 41, 42, 0.15);
-  display: flex;
-  flex-direction: column;
-  height: 473px;
-  overflow: hidden;
-  padding: 32px;
-  width: 657px;
-}
