--- client/src/pages/Landing/desktop.css
@@ -1,66 +1,0 @@
-.address {
-  left: 1299px;
-  letter-spacing: 0px;
-  line-height: 28px;
-  position: absolute;
-  top: 249px;
-  white-space: nowrap;
-  width: 68px;
-}
-.desktop-1 {
-  align-items: flex-start;
-  background-color: var(--white);
-  display: flex;
-  overflow: hidden;
-  width: 1440px;
-}
-.font-1 {
-  left: 1401px;
-  letter-spacing: 0px;
-  position: absolute;
-  top: 34px;
-}
-.overlap-group {
-  height: 1024px;
-  margin-left: -83px;
-  position: relative;
-  width: 1523px;
-}
-.span {
-  color: rgba(0, 0, 0, 0.44);
-  font-family: "Manrope", Helvetica;
-  font-size: 28px;
-  font-weight: 800;
-  line-height: 28px;
-}
-.span1 {
-  color: rgba(0, 0, 0, 1);
-  font-family: "Manrope", Helvetica;
-  font-size: 28px;
-  font-weight: 800;
-  line-height: 28px;
-}
-.text-1 {
-  left: 162px;
-  letter-spacing: 0px;
-  line-height: 98px;
-  position: absolute;
-  top: 127px;
-  width: 732px;
-}
-.text-3 {
-  left: 165px;
-  letter-spacing: 0px;
-  line-height: 40px;
-  position: absolute;
-  top: 369px;
-  width: 517px;
-}
-.untitled-design-8-1 {
-  height: 44px;
-  left: 1373px;
-  object-fit: cover;
-  position: absolute;
-  top: 240px;
-  width: 45px;
-}
